+++ .github/workflows/cli.yml
@@ -1,0 +1,35 @@
+name: Build and Install BankID CLI App
+
+on:
+    pull_request:
+    workflow_dispatch:
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+
+    steps:
+    - name: Checkout code
+      uses: actions/checkout@v2
+
+    - name: Set up Go
+      uses: actions/setup-go@v2
+      with:
+        go-version: '1.21'
+
+    - name: Build
+      run: go build -o bankid ./cmd/main
+
+    - name: Install
+      run: go install ./cmd/main
+      
+    - name: Build Docker Image
+      run: docker build -t bankid .
+
+    - name: Log in to GitHub Packages
+      run: echo ${{secrets.GH_TOKEN}} | docker login docker.pkg.github.com -u ${{secrets.USERNAME}} --password-stdin
+
+    - name: Push Docker Image
+      run: |
+            docker tag myapp docker.pkg.github.com/USERNAME/REPOSITORY/IMAGE_NAME:TAG
+            docker push docker.pkg.github.com/USERNAME/REPOSITORY/IMAGE_NAME:TAG
