+++ cmd/cli/main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"github.com/nicolaa5/bankid/cmd/cli"
+	"github.com/nicolaa5/bankid/cmd/cli/src"
 )
 
 func main() {
