+++ cmd/cli/src/auth.go
