+++ cmd/cli/src/sign.go
