+++ cmd/cli/src/terminal.go
